extracted optimizer creation into its own ingredient
