+++ tuned_lens/scripts/train_loop.py
@@ -6,10 +6,9 @@ from pathlib import Path
 from itertools import islice
 
 from simple_parsing import field
-from torch.distributed.optim import ZeroRedundancyOptimizer
 from torch.utils.data import DataLoader
 from tqdm.auto import tqdm
-from transformers import get_linear_schedule_with_warmup, PreTrainedModel
+from transformers import PreTrainedModel
 from tuned_lens import TunedLens
 from tuned_lens.residual_stream import ResidualStream
 from tuned_lens.utils import shift_labels, shift_preds, maybe_all_reduce
@@ -17,6 +16,7 @@ from tuned_lens.scripts.ingredients import (
     Model,
     Data,
     Distributed,
+    Optimizer,
 )
 import torch as th
 from dataclasses import dataclass
@@ -29,13 +29,6 @@ class LossChoice(enum.Enum):
     KL = "kl"
 
 
-class OptimizerOption(enum.Enum):
-    """Options for the optimizer to use when training the model."""
-
-    ADAM = "adam"
-    SGD = "sgd"
-
-
 @dataclass
 class Train:
     """Training loop for the tuned lens."""
@@ -45,6 +38,9 @@ class Train:
 
     data: Data
     """Data configuration."""
+
+    opt: Optimizer
+    """Optimizer configuration."""
 
     dist: Distributed
     """Configuration for how to distribute the training."""
@@ -57,20 +53,8 @@ class Train:
     constant: Optional[bool] = field(action="store_true")
     """Train only the bias term."""
 
-    lasso: float = 0.0
-    """LASSO (L1) regularization strength."""
-
-    lr_scale: float = 1.0
-    """The default LR (1e-3 for Adam, 1.0 for SGD) is scaled by this factor."""
-
-    momentum: float = 0.9
-    """Momentum coefficient for SGD, or beta1 for Adam."""
-
     num_steps: int = 250
     """Number of training steps."""
-
-    optimizer: OptimizerOption = OptimizerOption.SGD
-    """The type of optimizer to use."""
 
     output: Optional[Path] = field(alias=["-o"], default=None)
     """File to save the lenses to. Defaults to the model name."""
@@ -90,16 +74,6 @@ class Train:
     wandb: Optional[str] = None
     """Name of run in Weights & Biases."""
 
-    warmup_steps: Optional[int] = None
-    """Number of warmup steps. Defaults to min(0.1 * num_steps, 1000) for Adam and 0
-    for SGD."""
-
-    weight_decay: float = 1e-3
-    """Weight decay coefficient."""
-
-    zero: Optional[bool] = field(action="store_true")
-    """Use ZeroRedundancyOptimizer."""
-
     token_shift: Optional[int] = None
     """How to shift the labels wrt the input tokens (1 = next token, 0 = current token,
     -1 = previous token, etc.)"""
@@ -110,9 +84,7 @@ class Train:
     loss: LossChoice = LossChoice.KL
     """Loss function to use."""
 
-    def get_lens(
-        self, model: PreTrainedModel, must_use_cache: bool = False
-    ) -> TunedLens:
+    def get_lens(self, model: PreTrainedModel) -> TunedLens:
         """Load or create a TunedLens model."""
         if self.lens_name_or_path is None:
             lens = TunedLens(model)
@@ -155,17 +127,17 @@ class Train:
 
             # Approximate the true grad norm using the optimizer's moving
             # avg
-            corr = 1 - self.momentum**step
-            if self.optimizer == "sgd" and not self.zero:
+            corr = 1 - self.opt.momentum**step
+            if self.opt.optimizer == "sgd" and not self.opt.zero:
                 log_dict["grad_norm/" + name] = th.cat(
                     [
                         # Undo PyTorch's scaling of the gradient by
                         # 1 / (1 - β)
-                        (1 - self.momentum) * s["momentum_buffer"].flatten() / corr
+                        (1 - self.opt.momentum) * s["momentum_buffer"].flatten() / corr
                         for s in states
                     ]
                 ).norm()
-            elif self.optimizer == "adam" and not self.zero:
+            elif self.opt.optimizer == "adam" and not self.opt.zero:
                 log_dict["grad_norm/" + name] = th.cat(
                     [s["exp_avg"].flatten() / corr for s in states]
                 ).norm()
@@ -210,7 +182,7 @@ class Train:
 
     def execute(self):
         """Trains a TunedLens model against a transformer on a dataset."""
-        model = tokenizer = data = lens = None
+        model = tokenizer = data = lens = nats_to_bpb = None
         if self.dist.primary:
             model, tokenizer = self.model.load()
             data, nats_to_bpb = self.data.load(tokenizer)
@@ -221,13 +193,12 @@ class Train:
 
         self.dist.barrier()  # Wait for primary to finish loading
 
-        # TODO everyone else loads from cache
         if not self.dist.primary:
             model, tokenizer = self.model.load(must_use_cache=True)
             data, nats_to_bpb = self.data.load(tokenizer)
             lens = self.get_lens(model)
 
-        assert model and tokenizer and data and lens
+        assert model and tokenizer and data and lens and nats_to_bpb
 
         # Shard across GPUs
         model = self.dist.shard_model(model)
@@ -242,50 +213,13 @@ class Train:
         # Don't train the unembedding matrix or final layer norm
         params = [p for p in ddp_lens.parameters() if p.requires_grad]
 
-        β = self.momentum
-        if self.optimizer == "sgd":
-            config = dict(
-                # PyTorch's implementation effectively scales the LR by 1 / (1 - β),
-                # so we undo that here. See https://www.youtube.com/watch?v=k8fTYJPd3_I
-                # for discussion. Once we do this, the optimal LR seems to be unity.
-                lr=self.lr_scale * (1 - β),
-                momentum=β,
-                # Empirically Nesterov momentum seems to improve convergence speed.
-                nesterov=True,
-                weight_decay=self.weight_decay,
-            )
-            opt_class = th.optim.SGD
-        elif self.optimizer == "adam":
-            config = dict(
-                # Helps convergence slightly by ensuring that the LR actually decays
-                amsgrad=True,
-                betas=(β, 0.999),
-                lr=self.lr_scale * 1e-3,
-                weight_decay=self.weight_decay,
-            )
-            opt_class = th.optim.Adam
-        else:
-            raise ValueError(f"Unknown optimizer '{self.optimizer}'")
-
-        if self.zero:
-            opt = ZeroRedundancyOptimizer(params, optimizer_class=opt_class, **config)
-        else:
-            opt = opt_class(params, **config)  # type: ignore[call-arg]
+        opt = self.opt.create_optim(params)
 
         self.dist.barrier()
         print("All processes have completed setup. Starting training.")
 
-        if self.warmup_steps is None:
-            # Adam generally performs poorly without an LR warmup
-            if self.optimizer == "adam":
-                self.warmup_steps = min(1000, self.num_steps // 5)
-                print(f"Using {self.warmup_steps} LR warmup steps for Adam")
-            else:
-                self.warmup_steps = 0
+        scheduler = self.opt.create_scheduler(opt, self.num_steps)
 
-        scheduler = get_linear_schedule_with_warmup(
-            opt, self.warmup_steps, self.num_steps - self.warmup_steps
-        )
         if self.resume:
             assert self.resume.is_dir()
